BM-17320 fix: never remove user@domain.uid email alias
When BlueMind was migrated to using domain uid's segregated from domain names,
an alias was added to the domain, containing the domain uid.

In such cases, it was possible to remove the domain.uid alias from the domain.
When this happens, a hook took care of removing all user@domain.uid alises from
user emails.

This is bad because bluemind requires such an alias. Normally this is handled by
a UserSanitizer, but this is not called when the hook is called.

This modification just prevents from removing a user alias which is the domain.uid,
from a hook or not.

+++ parent/mailbackend/net.bluemind.backend.postfix/src/net/bluemind/backend/postfix/DomainHook.java
@@ -66,7 +66,9 @@ public class DomainHook extends DomainHookAdapter {
 		if (!deleted.isEmpty()) {
 			IInCoreMailboxes mailboxes = context.su().provider().instance(IInCoreMailboxes.class, domain.uid);
 			deleted.forEach(alias -> {
-				mailboxes.deleteEmailByAlias(alias);
+				if (!alias.equals(domain.uid)) {
+					mailboxes.deleteEmailByAlias(alias);
+				}
 			});
 		}
